requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
pypdf
python-dotenv
streamlit
faiss-cpu
sentence-transformers
langchain-huggingface